backend/requirements.txt
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
